backend/requirements.txt
requests==2.31.0
python-dotenv~=1.0.0
pydantic~=2.7.0
fastapi~=0.109.0
sqlalchemy~=2.0.25
environs~=9.5.0
uvicorn
alembic~=1.13.2
psycopg2-binary
# alembic init alembic
# alembic revision --autogenerate -m "create tables"
# alembic upgrade head

# uvicorn main:app --host 0.0.0.0 --port 8000
itsdangerous~=2.2.0
starlette~=0.36.3
librosa~=0.10.2.post1
numpy~=1.26.4
tensorflow~=2.17.0
pandas~=2.2.2
xgboost~=2.1.0
joblib~=1.4.2
pypdf
langchain
langchain_community
openai
langchain_openai
chromadb
langsmith
pyjwt
passlib[bcrypt]